refactor: split 6 oversized modules into 19 files (<300 LOC each)
Fixes #6

Split these modules to comply with the brick philosophy (<300 LOC):
- perspectives.rs (688 lines) -> 3 perspective files
- tools.rs (679 lines) -> definitions, executor, translate
- builder.rs (615 lines) -> binary, ghidra, source builders
- patterns.rs (399 lines) -> binary patterns, source patterns
- copilot.rs (386 lines) -> auth, client
- surface.rs (325 lines) -> binary surface, source sinks

No public API changes - parent mod.rs files re-export everything.
All 143 tests pass. Binary works identically.

Remaining files slightly over 300 (follow-up): ingest.rs (398),
sarif.rs (324), markdown.rs (316), parser.rs (314), mod.rs (305).
