+++ crates/core/src/agents/mod.rs
@@ -13,10 +13,13 @@ pub mod definition;
 pub mod discovery;
 pub mod pipeline;
 pub mod runner;
-pub mod tools;
+pub mod tool_definitions;
+pub mod tool_executor;
+pub mod tool_translate;
 
 pub use definition::{load_agent, AgentDefinition};
 pub use discovery::discover_agents;
 pub use pipeline::{default_pipeline, pipeline_from_names, AnalysisPipeline, PipelineStage};
 pub use runner::{AgentResult, AgentRunner};
-pub use tools::{agent_tools, execute_tool};
+pub use tool_definitions::{agent_tools, filter_tools};
+pub use tool_executor::execute_tool;
