+++ crates/core/src/agents/tool_executor.rs
@@ -1,153 +1,9 @@
-//! Tool definitions and executor for vulnerability assessment agents.
+//! Tool execution: dispatches tool calls against the real graph database.
 //!
-//! Each tool function executes actual SQL queries against the GraphDb,
-//! returning real data from ingested binaries. No fake data, no stubs.
+//! Every tool queries or mutates the actual database - no placeholder data.
 
 use crate::graph::GraphDb;
-use crate::llm::ToolDefinition;
-
-/// Return the full set of tool definitions that agents can call during analysis.
-pub fn agent_tools() -> Vec<ToolDefinition> {
-    vec![
-        ToolDefinition {
-            name: "query_graph".into(),
-            description: "Run a Cypher query against the code property graph and return results."
-                .into(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "cypher": {
-                        "type": "string",
-                        "description": "Cypher query to execute"
-                    }
-                },
-                "required": ["cypher"]
-            }),
-        },
-        ToolDefinition {
-            name: "read_function".into(),
-            description: "Read the decompiled or source code of a function by name or address."
-                .into(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "name": {
-                        "type": "string",
-                        "description": "Function name or address"
-                    }
-                },
-                "required": ["name"]
-            }),
-        },
-        ToolDefinition {
-            name: "get_callers".into(),
-            description: "Return all functions that call the specified function.".into(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "function": {
-                        "type": "string",
-                        "description": "Function name or address"
-                    }
-                },
-                "required": ["function"]
-            }),
-        },
-        ToolDefinition {
-            name: "get_callees".into(),
-            description: "Return all functions called by the specified function.".into(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "function": {
-                        "type": "string",
-                        "description": "Function name or address"
-                    }
-                },
-                "required": ["function"]
-            }),
-        },
-        ToolDefinition {
-            name: "lookup_cwe".into(),
-            description:
-                "Look up a CWE entry by ID and return its name, description, and mitigations."
-                    .into(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "cwe_id": {
-                        "type": "string",
-                        "description": "CWE identifier, e.g. CWE-787"
-                    }
-                },
-                "required": ["cwe_id"]
-            }),
-        },
-        ToolDefinition {
-            name: "create_finding".into(),
-            description: "Record a new vulnerability finding in the graph database.".into(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "title": {
-                        "type": "string",
-                        "description": "Short title for the finding"
-                    },
-                    "severity": {
-                        "type": "string",
-                        "enum": ["critical", "high", "medium", "low", "info"],
-                        "description": "Severity level"
-                    },
-                    "description": {
-                        "type": "string",
-                        "description": "Detailed description of the vulnerability"
-                    },
-                    "function": {
-                        "type": "string",
-                        "description": "Affected function name"
-                    },
-                    "cwe_id": {
-                        "type": "string",
-                        "description": "Associated CWE identifier"
-                    }
-                },
-                "required": ["title", "severity", "description"]
-            }),
-        },
-        ToolDefinition {
-            name: "search_similar".into(),
-            description:
-                "Search for code patterns similar to a given snippet using pattern matching.".into(),
-            parameters: serde_json::json!({
-                "type": "object",
-                "properties": {
-                    "code": {
-                        "type": "string",
-                        "description": "Code snippet to find similar patterns for"
-                    },
-                    "limit": {
-                        "type": "integer",
-                        "description": "Maximum number of results",
-                        "default": 10
-                    }
-                },
-                "required": ["code"]
-            }),
-        },
-    ]
-}
-
-/// Filter the full tool set to only tools listed in an agent's definition.
-pub fn filter_tools(all_tools: &[ToolDefinition], allowed: &[String]) -> Vec<ToolDefinition> {
-    if allowed.is_empty() {
-        return all_tools.to_vec();
-    }
-    all_tools
-        .iter()
-        .filter(|t| allowed.contains(&t.name))
-        .cloned()
-        .collect()
-}
+use super::tool_translate::{translate_to_sql, execute_read_query};
 
 /// Execute a single tool call against the real graph database.
 ///
@@ -165,8 +21,8 @@ pub fn execute_tool(
         "get_callers" => execute_get_callers(db, investigation_id, args),
         "get_callees" => execute_get_callees(db, investigation_id, args),
         "lookup_cwe" => execute_lookup_cwe(db, args),
-        "create_finding" => execute_create_finding(db, investigation_id, args),
-        "search_similar" => execute_search_similar(db, investigation_id, args),
+        "create_finding" => super::tool_translate::execute_create_finding(db, investigation_id, args),
+        "search_similar" => super::tool_translate::execute_search_similar(db, investigation_id, args),
         _ => {
             tracing::warn!("Unknown tool: {name}");
             Ok(serde_json::json!({
@@ -436,247 +292,6 @@ fn execute_lookup_cwe(
     }
 }
 
-/// Create a finding in the database.
-fn execute_create_finding(
-    db: &GraphDb,
-    investigation_id: &str,
-    args: &serde_json::Value,
-) -> anyhow::Result<serde_json::Value> {
-    let title = args
-        .get("title")
-        .and_then(|v| v.as_str())
-        .unwrap_or("Untitled");
-    let severity = args
-        .get("severity")
-        .and_then(|v| v.as_str())
-        .unwrap_or("medium");
-    let description = args
-        .get("description")
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-    let function = args
-        .get("function")
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-    let cwe_id = args
-        .get("cwe_id")
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-
-    let finding_id = uuid::Uuid::new_v4().to_string();
-    let timestamp = chrono::Utc::now().to_rfc3339();
-
-    tracing::info!("Tool create_finding: {title} [{severity}]");
-
-    let evidence = serde_json::json!({
-        "description": description,
-        "function": function,
-        "cwe_id": cwe_id,
-    });
-
-    db.execute(
-        "INSERT INTO findings (id, title, evidence, agent, timestamp, investigation_id, \
-         status, severity, category) \
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9)",
-        &[
-            &finding_id as &dyn rusqlite::types::ToSql,
-            &title,
-            &evidence.to_string(),
-            &"vuln_hunter",
-            &timestamp,
-            &investigation_id,
-            &"new",
-            &severity,
-            &cwe_id,
-        ],
-    )?;
-
-    Ok(serde_json::json!({
-        "status": "ok",
-        "finding_id": finding_id,
-        "title": title,
-        "severity": severity,
-        "investigation_id": investigation_id
-    }))
-}
-
-/// Search for functions with similar names or patterns.
-fn execute_search_similar(
-    db: &GraphDb,
-    investigation_id: &str,
-    args: &serde_json::Value,
-) -> anyhow::Result<serde_json::Value> {
-    let code = args
-        .get("code")
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-    let limit = args
-        .get("limit")
-        .and_then(|v| v.as_u64())
-        .unwrap_or(10) as usize;
-
-    let code_preview: String = code.chars().take(40).collect();
-    tracing::info!("Tool search_similar: {code_preview}...");
-
-    let search_pattern = format!("%{}%", code.replace('%', ""));
-
-    let mut stmt = db.conn().prepare(
-        "SELECT name, address, decompiled FROM functions \
-         WHERE investigation_id = ?1 AND (decompiled LIKE ?2 OR name LIKE ?2) \
-         LIMIT ?3",
-    )?;
-
-    let rows = stmt.query_map(
-        rusqlite::params![investigation_id, search_pattern, limit as i64],
-        |row| {
-            Ok((
-                row.get::<_, String>(0)?,
-                row.get::<_, String>(1)?,
-                row.get::<_, String>(2)?,
-            ))
-        },
-    )?;
-
-    let results: Vec<serde_json::Value> = rows
-        .filter_map(|r| match r {
-            Ok(v) => Some(v),
-            Err(e) => {
-                tracing::warn!("Error reading search result row: {e}");
-                None
-            }
-        })
-        .map(|(name, addr, decompiled)| {
-            let preview_raw = if decompiled.chars().count() > 200 {
-                let truncated: String = decompiled.chars().take(200).collect();
-                format!("{}...", truncated)
-            } else {
-                decompiled
-            };
-            let safe_preview = format!("<code_data>\n{}\n</code_data>", preview_raw);
-            serde_json::json!({
-                "name": name,
-                "address": addr,
-                "preview": safe_preview
-            })
-        })
-        .collect();
-
-    Ok(serde_json::json!({
-        "status": "ok",
-        "results": results,
-        "count": results.len()
-    }))
-}
-
-/// Translate common Cypher-like query patterns to parameterized SQL.
-///
-/// Returns `(sql, params)` where params contains the investigation_id
-/// for the `?1` placeholder. Only predefined patterns are supported;
-/// arbitrary SQL (including raw SELECT from the LLM) is rejected.
-fn translate_to_sql(query: &str, investigation_id: &str) -> Result<(String, Vec<String>), String> {
-    let q = query.trim();
-    let upper = q.to_uppercase();
-
-    if upper.contains("FUNCTION") && upper.contains("RETURN") {
-        return Ok((
-            "SELECT name, address, decompiled FROM functions WHERE investigation_id = ?1 LIMIT 50"
-                .to_string(),
-            vec![investigation_id.to_string()],
-        ));
-    }
-
-    if upper.contains("CALL") {
-        return Ok((
-            "SELECT f1.name as caller, f2.name as callee FROM calls c \
-             JOIN functions f1 ON c.caller_id = f1.id \
-             JOIN functions f2 ON c.callee_id = f2.id \
-             WHERE f1.investigation_id = ?1 LIMIT 50"
-                .to_string(),
-            vec![investigation_id.to_string()],
-        ));
-    }
-
-    if upper.contains("TAINT") || upper.contains("FLOW") {
-        return Ok((
-            "SELECT s.name, k.name, tf.path, tf.sanitized FROM taint_flows tf \
-             JOIN data_sources s ON tf.source_id = s.id \
-             JOIN data_sinks k ON tf.sink_id = k.id \
-             WHERE s.investigation_id = ?1 LIMIT 50"
-                .to_string(),
-            vec![investigation_id.to_string()],
-        ));
-    }
-
-    let q_preview: String = q.chars().take(80).collect();
-    Err(format!(
-        "Unsupported query pattern. Use Cypher-like queries with FUNCTION/RETURN, CALL, or TAINT/FLOW keywords. Got: {}",
-        q_preview
-    ))
-}
-
-/// Execute a read-only parameterized SQL query and return results as a Vec of JSON objects.
-fn execute_read_query(
-    db: &GraphDb,
-    sql: &str,
-    params: &[String],
-) -> anyhow::Result<Vec<serde_json::Value>> {
-    let stmt = db.conn().prepare(sql)?;
-    if !stmt.readonly() {
-        return Ok(vec![serde_json::json!({"error": "Only read-only queries are allowed. Write operations are not permitted."})]);
-    }
-    let mut stmt = stmt;
-    let column_count = stmt.column_count();
-    let column_names: Vec<String> = (0..column_count)
-        .map(|i| stmt.column_name(i).unwrap_or("?").to_string())
-        .collect();
-
-    let param_refs: Vec<&dyn rusqlite::types::ToSql> =
-        params.iter().map(|s| s as &dyn rusqlite::types::ToSql).collect();
-
-    let rows = stmt.query_map(param_refs.as_slice(), |row| {
-        let mut obj = serde_json::Map::new();
-        for (i, col_name) in column_names.iter().enumerate() {
-            let val: rusqlite::Result<String> = row.get(i);
-            match val {
-                Ok(s) => {
-                    obj.insert(col_name.clone(), serde_json::Value::String(s));
-                }
-                Err(_) => {
-                    let fval: rusqlite::Result<f64> = row.get(i);
-                    match fval {
-                        Ok(f) => {
-                            obj.insert(col_name.clone(), serde_json::json!(f));
-                        }
-                        Err(_) => {
-                            let ival: rusqlite::Result<i64> = row.get(i);
-                            match ival {
-                                Ok(n) => {
-                                    obj.insert(col_name.clone(), serde_json::json!(n));
-                                }
-                                Err(_) => {
-                                    obj.insert(col_name.clone(), serde_json::Value::Null);
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Ok(serde_json::Value::Object(obj))
-    })?;
-
-    let results: Vec<serde_json::Value> = rows
-        .filter_map(|r| match r {
-            Ok(v) => Some(v),
-            Err(e) => {
-                tracing::warn!("Error reading query result row: {e}");
-                None
-            }
-        })
-        .collect();
-    Ok(results)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -892,22 +507,6 @@ mod tests {
     }
 
     #[test]
-    fn test_filter_tools() {
-        let all = agent_tools();
-        let filtered = filter_tools(&all, &["query_graph".into(), "read_function".into()]);
-        assert_eq!(filtered.len(), 2);
-        assert!(filtered.iter().any(|t| t.name == "query_graph"));
-        assert!(filtered.iter().any(|t| t.name == "read_function"));
-    }
-
-    #[test]
-    fn test_filter_tools_empty_allows_all() {
-        let all = agent_tools();
-        let filtered = filter_tools(&all, &[]);
-        assert_eq!(filtered.len(), all.len());
-    }
-
-    #[test]
     fn test_query_graph_rejects_destructive_query() {
         let db = GraphDb::in_memory().unwrap();
         let inv_id = "test-inv";
@@ -931,59 +530,5 @@ mod tests {
             )
             .unwrap();
         assert_eq!(count, 0, "Destructive query must not modify the database");
-    }
-
-    #[test]
-    fn test_readonly_enforcement_in_execute_read_query() {
-        let db = GraphDb::in_memory().unwrap();
-
-        // Directly test execute_read_query with a write statement
-        let result = execute_read_query(&db, "INSERT INTO functions (id, name) VALUES ('evil', 'injected')", &[]).unwrap();
-
-        // Should return an error entry instead of executing
-        assert_eq!(result.len(), 1);
-        assert!(
-            result[0].get("error").is_some(),
-            "Write query should be rejected by readonly check"
-        );
-
-        // Verify nothing was inserted
-        let count: i64 = db
-            .conn()
-            .query_row(
-                "SELECT count(*) FROM functions WHERE name = 'injected'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(count, 0);
-    }
-
-    #[test]
-    fn test_translate_to_sql_blocks_unrecognised_patterns() {
-        // INSERT should be rejected
-        let result = translate_to_sql("INSERT INTO functions VALUES ('x', 'y')", "inv1");
-        assert!(result.is_err());
-
-        // UPDATE should be rejected
-        let result = translate_to_sql("UPDATE functions SET name = 'hacked'", "inv1");
-        assert!(result.is_err());
-
-        // Raw SELECT should also be rejected (no pass-through)
-        let result = translate_to_sql("SELECT * FROM functions", "inv1");
-        assert!(result.is_err());
-
-        // Recognised Cypher-like patterns should work and use parameterized queries
-        let result = translate_to_sql("MATCH (f:Function) RETURN f", "inv1");
-        assert!(result.is_ok());
-        let (sql, params) = result.unwrap();
-        assert!(sql.contains("?1"));
-        assert_eq!(params, vec!["inv1"]);
-
-        let result = translate_to_sql("MATCH (c:Call) RETURN c", "inv1");
-        assert!(result.is_ok());
-        let (sql, params) = result.unwrap();
-        assert!(sql.contains("?1"));
-        assert_eq!(params, vec!["inv1"]);
     }
 }
