+++ crates/core/src/analysis/mod.rs
@@ -6,17 +6,27 @@ pub mod findings;
 pub mod hardening;
 pub mod orchestrator;
 pub mod patterns;
-pub mod perspectives;
+pub mod patterns_binary;
+pub mod patterns_source;
+pub mod perspective_context;
+pub mod perspective_dataflow;
+pub mod perspective_pattern;
 pub mod semgrep;
 pub mod severity;
 pub mod surface;
+pub mod surface_binary;
+pub mod surface_source;
 pub mod taint;
 pub mod variant;
 
 pub use findings::{Finding, FindingLocation, FindingStatus, FindingUpdate};
 pub use hardening::format_hardening;
 pub use orchestrator::{AnalysisCycle, AnalysisOrchestrator};
-pub use patterns::{DangerousApiDetector, DangerousApiHit, DangerCategory, Severity};
+pub use patterns::{DangerCategory, DangerousApiHit, Severity};
+pub use patterns_binary::DangerousApiDetector;
+pub use perspective_context::context_perspective;
+pub use perspective_dataflow::dataflow_perspective;
+pub use perspective_pattern::pattern_perspective;
 pub use semgrep::SemgrepRunner;
 pub use severity::compute_severity;
 pub use surface::{
