+++ crates/core/src/graph/mod.rs
@@ -4,10 +4,13 @@
 //! common query helpers.
 
 pub mod builder;
+pub mod builder_binary;
+pub mod builder_ghidra;
+pub mod builder_source;
 pub mod db;
 pub mod queries;
 pub mod types;
 
-pub use builder::{GraphBuilder, InsertCounts, SourceInsertCounts};
+pub use builder::{GraphBuilder, GhidraInsertCounts, InsertCounts, SourceInsertCounts};
 pub use db::GraphDb;
 pub use types::{NodeLabel, RelationshipType};
